# perf: Optimize regular numbers parse logics