@filzrev 2024-10-09
### Microbenchmark results

`Optimized` version run about 10x faster. and memory allocation is reduced about 50%.

**Baseline**

| Method      | Job                         | Runtime            | Mean     | Error    | StdDev  | Exceptions | Gen0   | Gen1   | Allocated |
|------------ |---------------------------- |------------------- |---------:|---------:|--------:|-----------:|-------:|-------:|----------:|
| Deserialize | ShortRun-.NET 8.0           | .NET 8.0           | 176.9 us | 18.02 us | 0.99 us |    18.0000 | 7.3242 |      - |  29.96 KB |
| Deserialize | ShortRun-.NET Framework 4.7 | .NET Framework 4.7 | 588.5 us | 60.13 us | 3.30 us |    23.0000 | 9.7656 | 0.9766 |  43.63 KB |

**Optimized**
| Method      | Job                         | Runtime            | Mean     | Error    | StdDev   | Exceptions | Gen0   | Gen1   | Allocated |
|------------ |---------------------------- |------------------- |---------:|---------:|---------:|-----------:|-------:|-------:|----------:|
| Deserialize | ShortRun-.NET 8.0           | .NET 8.0           | 20.88 us | 3.234 us | 0.177 us |          - | 4.0283 |      - |  16.45 KB |
| Deserialize | ShortRun-.NET Framework 4.7 | .NET Framework 4.7 | 52.78 us | 9.504 us | 0.521 us |          - | 5.3711 | 0.1221 |  23.66 KB |

**Test Code**
```csharp
[MemoryDiagnoser]
[ExceptionDiagnoser]
[ShortRunJob(RuntimeMoniker.Net80)]
[ShortRunJob(RuntimeMoniker.Net47)]
public class AttemptUnknownTypeDeserializationBenchmarks
{
    private static readonly IDeserializer deserializer = new DeserializerBuilder().WithAttemptingUnquotedStringTypeDeserialization().Build();

    private static readonly string TestData =
        """
        byteMaxValue:   255
        shortMaxValue:  32767
        longMaxValue :  9223372036854775807
        ulongMaxValue:  18446744073709551615
        floatMaxValue:  3.4028235E+38
        doubleMaxValue: 1.7976931348623157E+308
        """;

    [Benchmark]
    public object? Deserialize()
    {
        return deserializer.Deserialize(TestData);
    }
}

```