@EdwardCooke 2024-10-09
I like performance fixes. And this is a pretty big one! If you can address the 2 comments then I'll merge this in and get it out.