@filzrev 2024-10-10
Apart from this PR.
I've noticed another performance relating issue.

https://github.com/aaubry/YamlDotNet/blob/master/YamlDotNet/Serialization/NodeDeserializers/ScalarNodeDeserializer.cs#L383

Above regex checking `regular number`(integer or floating number).
But `any text that contains digits` match regex.
So unneeded `Parse` operations are called multiple times.

Is it able to change Regex to use more strict regex with `^` and `$` ?
or is there any YAML specs that require text partial match?