@EdwardCooke 2024-10-10
You can add those to that regex without any issue I would think.