@filzrev 2024-10-10
> You can add those to that regex without any issue I would think.

Thanks for your response.
I'll try to create another PR after this PR is merged.