@filzrev 2024-10-26
I've removed wrong comment in commit.

`float.IsNormal` is available .NET Core 2.1 and The behavior will be different from what I'm expecting.
https://learn.microsoft.com/en-us/dotnet/api/system.single.isnormal?view=net-8.0#remarks